Only move file if not symlink
[Fixes Issue #1]

+++ makesymlinks.sh
@@ -26,7 +26,12 @@ echo "...done"
 # move any existing dotfiles in homedir to dotfiles_old directory, then create
 # symlinks
 for file in $files; do
-  echo "Moving any existing dotfiles from ~ to $olddir"
+  if [ -h $file ] # if the file is already a symbolic link, skip
+  then
+    break
+  fi
+
+  echo "Moving $file from ~ to $olddir"
   mv ~/$file ~/.dotfiles_old/
   echo "Creating symlink to $file in home directory."
   ln -s $dir/$file ~/$file
